+++ src/Dynatrace.OpenKit/Core/Objects/Session.cs
@@ -201,12 +201,6 @@ namespace Dynatrace.OpenKit.Core.Objects
 
         public void IdentifyUser(string userTag)
         {
-            if (string.IsNullOrEmpty(userTag))
-            {
-                logger.Warn($"{this} IdentifyUser: userTag must not be null or empty");
-                return;
-            }
-
             if (logger.IsDebugEnabled)
             {
                 logger.Debug($"{this} IdentifyUser({userTag})");